refactor: moved qcu and pairing form back to their own components
