+++ src/widgets/component-new-edit/components/response-format/single/response-format-single.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
@@ -6,33 +6,27 @@ import { Field, FormSection, change, formValueSelector } from 'redux-form';
 
 import {
   CODES_LIST_INPUT_ENUM,
-  COMPONENT_TYPE,
   DATATYPE_VIS_HINT,
   DEFAULT_CODES_LIST_SELECTOR_PATH,
   DEFAULT_FORM_NAME,
-  DIMENSION_FORMATS,
   QUESTION_TYPE_ENUM,
 } from '../../../../../constants/pogues-constants';
 import GenericOption from '../../../../../forms/controls/generic-option';
 import ListRadios from '../../../../../forms/controls/list-radios';
-import Select from '../../../../../forms/controls/select';
 import Dictionary from '../../../../../utils/dictionary/dictionary';
 import { getCurrentSelectorPath } from '../../../../../utils/widget-utils';
 import { CodesLists } from '../../../../codes-lists';
 import SuggesterLists from '../../../../codes-lists/containers/suggester-lists-container';
 
-const { SINGLE_CHOICE, PAIRING, TABLE } = QUESTION_TYPE_ENUM;
+const { SINGLE_CHOICE } = QUESTION_TYPE_ENUM;
 const { CHECKBOX, RADIO, DROPDOWN, SUGGESTER } = DATATYPE_VIS_HINT;
-const { QUESTION, LOOP } = COMPONENT_TYPE;
-const { LIST } = DIMENSION_FORMATS;
 const { REF } = CODES_LIST_INPUT_ENUM;
 
+/** Form to create a QCU. */
 function ResponseFormatSingle({
   selectorPathParent,
   responseFormatType,
   showMandatory,
-  componentsStore,
-  collectedVariablesStore,
   visHint,
   path,
   formName,
@@ -49,72 +43,6 @@ function ResponseFormatSingle({
     ? `${selectorPathParent}.${selectorPath}`
     : selectorPath;
 
-  // The sequences directly depending of the loops
-  const loopSequences = Object.values(componentsStore)
-    .filter((component) => component.type === LOOP)
-    .reduce((acc, loop) => {
-      return [
-        ...acc,
-        Object.values(componentsStore)
-          .filter(
-            // members are all SEQUENCE or all SUBSEQUENCE
-            (component) =>
-              component.type === componentsStore[loop.initialMember].type &&
-              component.weight >= componentsStore[loop.initialMember].weight &&
-              component.weight <= componentsStore[loop.finalMember].weight,
-          )
-          .map((component) => component.id),
-      ];
-    }, [])
-    .flat();
-
-  // gets the ids of the responses of dynamic arrays and questions in loops
-  const RosterVariablesId = Object.values(componentsStore)
-    // keep only component (remove root questionnaire, i.e the component where parent is equal to "")
-    .filter(
-      (component) => component.parent !== '' && component.parent !== undefined,
-    )
-    .filter(
-      (component) =>
-        component.type === QUESTION &&
-        // dynamic arrays
-        ((component.responseFormat.type === TABLE &&
-          component.responseFormat.TABLE.PRIMARY.type === LIST) ||
-          // questions directly depending on loop loopSequences
-          loopSequences.includes(component.parent) ||
-          // questions indirectly depending on loop loopSequences
-          loopSequences.includes(componentsStore[component.parent]?.parent)),
-    )
-    .map((q) => q.collectedVariables)
-    .flat();
-
-  // reduces from each question and from each question's variable the pairing structure including the id and the name of the RosterVariablesId
-  const pairingSourceVariable = Object.values(collectedVariablesStore)
-    .filter((question) => typeof question === 'object')
-    .reduce(
-      (acc, questionVariable) => [
-        ...acc,
-        Object.values(questionVariable)
-          .reduce(
-            (acc2, collectedVariable) =>
-              RosterVariablesId.includes(collectedVariable.id)
-                ? [
-                    ...acc2,
-                    <GenericOption
-                      key={collectedVariable.id}
-                      value={collectedVariable.id}
-                    >
-                      {collectedVariable.name}
-                    </GenericOption>,
-                  ]
-                : acc2,
-            [],
-          )
-          .flat(),
-      ],
-      [],
-    );
-
   useEffect(() => {
     if (visHint === SUGGESTER) {
       change(formName, `${path}${DEFAULT_CODES_LIST_SELECTOR_PATH}.panel`, REF);
@@ -129,55 +57,37 @@ function ResponseFormatSingle({
 
   return (
     <FormSection name={selectorPath} className="response-format__single">
-      {responseFormatType === PAIRING ? (
-        <Field
-          name="scope"
-          component={Select}
-          label={Dictionary.pairingSourceVariable}
-          required
-        >
-          <GenericOption key="" value="">
-            {Dictionary.selectBasedOn}
-          </GenericOption>
-          {pairingSourceVariable}
-        </Field>
-      ) : (
-        <>
-          <Field
-            name="visHint"
-            component={ListRadios}
-            label={Dictionary.visHint}
-            required
-          >
-            <GenericOption key={RADIO} value={RADIO}>
-              {Dictionary.radio}
-            </GenericOption>
-            <GenericOption key={DROPDOWN} value={DROPDOWN}>
-              {Dictionary.dropdown}
-            </GenericOption>
-            <GenericOption key={CHECKBOX} value={CHECKBOX}>
-              {Dictionary.checkbox}
-            </GenericOption>
-            <GenericOption key={SUGGESTER} value={SUGGESTER}>
-              {Dictionary.suggester}
-            </GenericOption>
-          </Field>
-          {visHint !== SUGGESTER && (
-            <div className="ctrl-checkbox" style={styleMandatory}>
-              <label htmlFor="rf-single-mandatory">
-                {Dictionary.mandatory}
-              </label>
-              <div>
-                <Field
-                  name="mandatory"
-                  id="rf-single-mandatory"
-                  component="input"
-                  type="checkbox"
-                />
-              </div>
-            </div>
-          )}
-        </>
+      <Field
+        name="visHint"
+        component={ListRadios}
+        label={Dictionary.visHint}
+        required
+      >
+        <GenericOption key={RADIO} value={RADIO}>
+          {Dictionary.radio}
+        </GenericOption>
+        <GenericOption key={DROPDOWN} value={DROPDOWN}>
+          {Dictionary.dropdown}
+        </GenericOption>
+        <GenericOption key={CHECKBOX} value={CHECKBOX}>
+          {Dictionary.checkbox}
+        </GenericOption>
+        <GenericOption key={SUGGESTER} value={SUGGESTER}>
+          {Dictionary.suggester}
+        </GenericOption>
+      </Field>
+      {visHint !== SUGGESTER && (
+        <div className="ctrl-checkbox" style={styleMandatory}>
+          <label htmlFor="rf-single-mandatory">{Dictionary.mandatory}</label>
+          <div>
+            <Field
+              name="mandatory"
+              id="rf-single-mandatory"
+              component="input"
+              type="checkbox"
+            />
+          </div>
+        </div>
       )}
       {visHint === SUGGESTER ? (
         <>
@@ -213,8 +123,6 @@ ResponseFormatSingle.propTypes = {
   selectorPathParent: PropTypes.string,
   responseFormatType: PropTypes.string,
   showMandatory: PropTypes.bool,
-  componentsStore: PropTypes.object,
-  collectedVariablesStore: PropTypes.object,
   visHint: PropTypes.string,
   path: PropTypes.string,
   formName: PropTypes.string,
@@ -227,8 +135,6 @@ ResponseFormatSingle.defaultProps = {
   selectorPathParent: undefined,
   responseFormatType: SINGLE_CHOICE,
   showMandatory: true,
-  componentsStore: {},
-  collectedVariablesStore: {},
   visHint: undefined,
   path: SINGLE_CHOICE,
   formName: DEFAULT_FORM_NAME,
@@ -243,8 +149,6 @@ const mapStateToProps = (state, { selectorPathParent, responseFormatType }) => {
     responseFormatType ?? SINGLE_CHOICE
   }.`;
   return {
-    componentsStore: state.appState.activeComponentsById,
-    collectedVariablesStore: state.appState.collectedVariableByQuestion,
     visHint: selector(state, `${path}visHint`),
     allowArbitraryResponse: selector(state, `${path}allowArbitraryResponse`),
     path,
